Remove styled arrow component in favour of inline styling
Removed the styled arrow component because tailwind had issues geting the positioning correct.
i've attempted to use `styled-components` to get the position of the arrow absolutely perfectly, but i ran into an issue with it and typescript. it worked locally but wouldn't pass the typescript checker citing:
"This expression is not callable.". i attempted to fix it by re-installing react types again though that didn't help.
instead of spending too much time debugging this issue, i didn't need to do a lot of styling, so i simply added the needed positioning styled inline and called it a day