Move global css import to layout
Moved the import of the global css file into the layout component from the gatsby-browser config in an attempt to make the css load faster and not flash unstyled content.

i don't percieve a difference, unstyled content still flashes. i'll keep it as is in the hopes that once it's built, it won't actually flash. so i'll revisit this issue again when i'm done with the site and ready to deploy