Fix bg colors in production
Changed how skill tags are dynamically assigned background colors. originally, it was done so:

`${p => "bg-"+p.color+"-600"}`

While it worked locally, it introduced an issue when the project was built for production. because the full background class name was not mentioned statically, it got automatically pruned from the source files by mistake. now the class name is not interpolated and is implemented as a `switch/case` which is uglier but does not trigger pruning so now the colors are displayed correctly in production builds